refactor: split client components into separate barrel
