+++ src/components/chat/chat-interface.tsx
@@ -38,7 +38,7 @@ import { CONSTANTS } from './constants'
 import { useDocumentUploader } from './document-uploader'
 import { useChatState } from './hooks/use-chat-state'
 import { useCustomSystemPrompt } from './hooks/use-custom-system-prompt'
-import { initializeRenderers } from './renderers'
+import { initializeRenderers } from './renderers/client'
 import type { VerificationState } from './types'
 // Lazy-load heavy, non-critical UI to reduce initial bundle and speed up FCP
 const VerifierSidebarLazy = dynamic(
