+++ src/components/chat/renderers/client.ts
@@ -1,0 +1,28 @@
+'use client'
+
+// Client-only exports for React components
+// Import from this file when using these components in client components
+
+// Registry functions (needed by client components)
+export { getRendererRegistry } from './registry'
+
+// Default renderers
+export { DefaultInputRenderer } from './default/DefaultInputRenderer'
+export { DefaultMessageRenderer } from './default/DefaultMessageRenderer'
+
+// Components
+export { DocumentList } from './components/DocumentList'
+export { MessageActions } from './components/MessageActions'
+export { MessageContent } from './components/MessageContent'
+export { ThoughtProcess } from './components/ThoughtProcess'
+
+// Initialization function that uses client components
+import { DefaultInputRenderer } from './default/DefaultInputRenderer'
+import { DefaultMessageRenderer } from './default/DefaultMessageRenderer'
+import { getRendererRegistry } from './registry'
+
+export function initializeRenderers(): void {
+  const registry = getRendererRegistry()
+  registry.setDefaultMessageRenderer(DefaultMessageRenderer)
+  registry.setDefaultInputRenderer(DefaultInputRenderer)
+}
